# Define available beatmap attributes