--- a/app/Models/Beatmap.php
+++ b/app/Models/Beatmap.php
@@ -39,6 +39,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property int $passcount
  * @property int $playcount
  * @property int $playmode
+ * @property int $score_version
  * @property int $total_length
  * @property int $user_id
  * @property string $version