--- a/app/Transformers/BeatmapCompactTransformer.php
+++ b/app/Transformers/BeatmapCompactTransformer.php
@@ -21,17 +21,15 @@ class BeatmapCompactTransformer extends TransformerAbstract
 
     public function transform(Beatmap $beatmap)
     {
-        $attrs = $beatmap->getAttributes();
-
         return [
-            'beatmapset_id' => $attrs['beatmapset_id'] ?? null,
-            'difficulty_rating' => $beatmap->getDifficultyRatingAttribute(),
-            'id' => $attrs['beatmap_id'] ?? null,
-            'mode' => $beatmap->getModeAttribute(),
+            'beatmapset_id' => $beatmap->beatmapset_id,
+            'difficulty_rating' => $beatmap->difficultyrating,
+            'id' => $beatmap->beatmap_id,
+            'mode' => $beatmap->mode,
             'status' => $beatmap->status(),
-            'total_length' => $attrs['total_length'] ?? null,
-            'user_id' => $attrs['user_id'] ?? null,
-            'version' => $beatmap->getVersionAttribute(),
+            'total_length' => $beatmap->total_length,
+            'user_id' => $beatmap->user_id,
+            'version' => $beatmap->version,
         ];
     }