--- a/app/Transformers/BeatmapTransformer.php
+++ b/app/Transformers/BeatmapTransformer.php
@@ -18,29 +18,25 @@ class BeatmapTransformer extends BeatmapCompactTransformer
 
     public function transform(Beatmap $beatmap)
     {
-        $result = parent::transform($beatmap);
-
-        $attrs = $beatmap->getAttributes();
-
-        return array_merge($result, [
-            'accuracy' => $attrs['diff_overall'] ?? null,
-            'ar' => $attrs['diff_approach'] ?? null,
-            'bpm' => $attrs['bpm'] ?? null,
+        return array_merge(parent::transform($beatmap), [
+            'accuracy' => $beatmap->diff_overall,
+            'ar' => $beatmap->diff_approach,
+            'bpm' => $beatmap->bpm,
             'convert' => $beatmap->convert,
-            'count_circles' => $attrs['countNormal'] ?? null,
-            'count_sliders' => $attrs['countSlider'] ?? null,
-            'count_spinners' => $attrs['countSpinner'] ?? null,
-            'cs' => $beatmap->getDiffSizeAttribute(),
-            'deleted_at' => json_time_from_db_timestamp($attrs['deleted_at'] ?? null),
-            'drain' => $attrs['diff_drain'] ?? null,
-            'hit_length' => $attrs['hit_length'] ?? null,
+            'count_circles' => $beatmap->countNormal,
+            'count_sliders' => $beatmap->countSlider,
+            'count_spinners' => $beatmap->countSpinner,
+            'cs' => $beatmap->diff_size,
+            'deleted_at' => $beatmap->deleted_at_json,
+            'drain' => $beatmap->diff_drain,
+            'hit_length' => $beatmap->hit_length,
             'is_scoreable' => $beatmap->isScoreable(),
-            'last_updated' => json_time_from_db_timestamp($attrs['last_update'] ?? null),
-            'mode_int' => $attrs['playmode'] ?? null,
-            'passcount' => $attrs['passcount'] ?? null,
-            'playcount' => $attrs['playcount'] ?? null,
-            'ranked' => $attrs['approved'] ?? null,
-            'url' => route('beatmaps.show', ['beatmap' => $attrs['beatmap_id'] ?? null]),
+            'last_updated' => $beatmap->last_update_json,
+            'mode_int' => $beatmap->playmode,
+            'passcount' => $beatmap->passcount,
+            'playcount' => $beatmap->playcount,
+            'ranked' => $beatmap->approved,
+            'url' => route('beatmaps.show', ['beatmap' => $beatmap->getKey()]),
         ]);
     }
 }