@peppy 2022-10-04
Would be helpful if you can include any profiling results in the PR description.